edit api

+++ src/components/Table.jsx
@@ -1,12 +1,38 @@
 import { Space, Table, Tag } from 'antd';
-import { Button } from 'antd';
+import { Button ,Modal,Input} from 'antd';
+import { useState } from 'react';
 
 
 export const MyTable = (props) => {
 
+
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [editState,setEditState]=useState({
+    name:""
+
+  })
+  const showModal = (record) => {
+    console.log(record);
+    setEditState(record);
+    setIsModalOpen(true);
+  };
+  const handleOk = () => {
+    props.editEmployee(editState);
+    setIsModalOpen(false);
+  };
+  const handleCancel = () => {
+    setIsModalOpen(false);
+  };
+
+
   const deleteData=(id)=>{
     props.deleteData(id);
    // console.log(id);
+
+  }
+
+  const handleChange=(event)=>{
+    setEditState({...editState,[event.target.name]:event.target.value})
 
   }
 
@@ -35,7 +61,8 @@ export const MyTable = (props) => {
         
 
         <Space size="middle">
-          <a>Invite {record.name}</a>
+        <Button type="primary" onClick={()=>showModal(record)}>Edit</Button>
+
           <Button type="primary" onClick={()=>deleteData(record.id)}>Delete</Button>
   
         </Space>
@@ -44,7 +71,14 @@ export const MyTable = (props) => {
   ];
 
 return (
-<Table columns={columns} dataSource={props.data} />
+  <>
+      <Modal title="Basic Modal" open={isModalOpen} onOk={handleOk} onCancel={handleCancel}>
+      <Input placeholder="Name" value={editState.name} name="name" onChange={handleChange} />
+      </Modal>
+
+      <Table columns={columns} dataSource={props.data} />
+  </>
+
 )
